requirements.txt
google.generativeai
flask
gtts
speechrecognition